pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finrag"
version = "0.1.0"
description = "Production-grade Financial Reports RAG Analyzer for SEC 10-K filings"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # --- Core Framework ---
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-index-retrievers-bm25>=0.1.0",

    # --- Agentic Orchestration ---
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",

    # --- LLM Backends ---
    "openai>=1.30.0",

    # --- Embeddings ---
    "sentence-transformers>=3.0.0",

    # --- Vector Store ---
    "chromadb>=0.5.0",

    # --- Config & Validation ---
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # --- CLI ---
    "typer>=0.12.0",
    "rich>=13.7.0",

    # --- HTTP / Download ---
    "httpx>=0.27.0",
    "tenacity>=8.3.0",

    # --- Document Parsing ---
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "pandas>=2.2.0",

    # --- API Server (Day 11) ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",

    # --- Observability ---
    "structlog>=24.2.0",

    # --- Utilities ---
    "python-ulid>=2.2.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.4.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-beautifulsoup4>=4.12.0",
]
llamaparse = [
    # Optional: LlamaParse for superior table extraction (requires API key)
    "llama-parse>=0.4.0",
]

[project.scripts]
finrag = "ops.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# ── Black ──────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

# ── Ruff ───────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # line length handled by black

# ── MyPy ───────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ── Pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"